+++ controllers/indexController.js
@@ -3,7 +3,6 @@ const asyncHandler = require("express-async-handler");
 const indexController = {
   getIndex: asyncHandler((req, res) => {
     console.log("getIndex running...");
-    console.log("req.url", req.url);
     res.render("index");
   }),
 };