+++ validators/validators.js
@@ -2,6 +2,7 @@ const validateCreateAccount = require("./createAccountValidator");
 const validateFilename = require("./filenameValidator");
 const validateFolder = require("./folderValidator");
 const validateLogin = require("./loginValidator");
+const { validateParams } = require("./paramsValidator");
 const validateShareDuration = require("./shareDurationValidator");
 const validateUpload = require("./uploadValidator");
 
@@ -10,6 +11,7 @@ module.exports = {
   validateFilename,
   validateFolder,
   validateLogin,
+  validateParams,
   validateShareDuration,
   validateUpload,
 };